Revert "currentTIme类型改为double"
This reverts commit 98c0f3c0dbe7c0d596e59547761e059d98c71a6c.

+++ src/main/java/com/empty/entity/DanEntity.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class DanEntity {
 	private Integer danId;
 	private Date danSendTime;
-	private Double danCurrTime;	//弹幕弹出时间单位是秒 和myVideoCurrentTime一样
+	private Integer danCurrTime;	//弹幕弹出时间单位是秒 和myVideoCurrentTime一样
 	private String danContent;
 	private String danStyle;	//这里可以直接存放css json对象
 	private Integer videoId;
@@ -59,11 +59,11 @@ public class DanEntity {
 		this.userId = userId;
 	}
 
-	public Double getDanCurrTime() {
+	public Integer getDanCurrTime() {
 		return danCurrTime;
 	}
 
-	public void setDanCurrTime(Double danCurrTime) {
+	public void setDanCurrTime(Integer danCurrTime) {
 		this.danCurrTime = danCurrTime;
 	}
 }
